Move issues tests to correct file
