+++ client/test/app/intake/util/issues.test.js
@@ -1,4 +1,5 @@
-import { formatAddedIssues } from 'app/intake/util/issues';
+import { formatAddedIssues, getAddIssuesFields } from 'app/intake/util/issues';
+import { getAddIssuesFieldsSamples } from '../../../data/intake/intakes';
 
 const addedIssues = [
   {
@@ -105,6 +106,21 @@ const expected = [
   }
 ];
 
+const testVeteran = 'Bob Something (000000001)';
+
+const intakesCases = {
+  docketTypeNotHearing: 0,
+  hearingTypeNull: 1,
+  hearingTypePresent: 2,
+};
+
+const hearingTypeIsPresent = (fields) => {
+  // findIndex returns -1 if an entry with field = 'hearing type' isn't found.
+  let idx = fields.findIndex((entry) => entry.field.toLowerCase() === 'hearing type');
+
+  return idx !== -1;
+};
+
 describe('formatAddedIssues', () => {
   describe('nonrating issues', () => {
     describe('adding new', () => {
@@ -134,3 +150,36 @@ describe('formatAddedIssues', () => {
     });
   });
 });
+
+describe('getAddIssueFields - form_type: appeal - Hearing type field', () => {
+  it('is not present because docket type is not hearing', async () => {
+    const fields = getAddIssuesFields(
+      'appeal',
+      testVeteran,
+      getAddIssuesFieldsSamples[intakesCases.docketTypeNotHearing]
+    );
+
+    expect(hearingTypeIsPresent(fields)).toBe(false);
+  });
+
+  it('is not present because hearing type is null despite docket type being hearing', async () => {
+    const fields = getAddIssuesFields(
+      'appeal',
+      testVeteran,
+      getAddIssuesFieldsSamples[intakesCases.hearingTypeNull]
+    );
+
+    expect(hearingTypeIsPresent(fields)).toBe(false);
+  });
+
+  it('is present because docket type is hearing and hearing type is not null', async () => {
+    const fields = getAddIssuesFields(
+      'appeal',
+      testVeteran,
+      getAddIssuesFieldsSamples[intakesCases.hearingTypePresent]
+    );
+
+    expect(hearingTypeIsPresent(fields)).toBe(true);
+  });
+});
+
